update: refactor global of Dagviz package
